Implementacao da RN002 e RN003 - O campo codigo deve ser autogerado.

+++ src/main/java/com/supermercado/repository/ClienteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.supermercado.domain.Cliente;
 
-public interface ClienteRepository extends MongoRepository<Cliente, Integer> {
+public interface ClienteRepository extends MongoRepository<Cliente, Long> {
 
 }
